Made it work for sample data
Changed it from using the fake data that I invented to instead use the data from the course, and updated the app js file so it now has the top bar be used as a file upload section, with options for typing name or select (proof of concept?). Also switched example text for explanatory text on the functionality. - N. A.